Disallow removal of entries from readonly archives

+++ Sources/ZIPFoundation/Archive+Writing.swift
@@ -150,13 +150,13 @@ extension Archive {
     ///   - progress: A progress object that can be used to track or cancel the remove operation.
     /// - Throws: An error if the `Entry` is malformed or the receiver is not writable.
     public func remove(_ entry: Entry, bufferSize: UInt32 = defaultReadChunkSize, progress: Progress? = nil) throws {
+        guard self.accessMode != .read else { throw ArchiveError.unwritableArchive }
         let manager = FileManager()
         let tempDir = self.uniqueTemporaryDirectoryURL()
         defer { try? manager.removeItem(at: tempDir) }
         let uniqueString = ProcessInfo.processInfo.globallyUniqueString
         let tempArchiveURL =  tempDir.appendingPathComponent(uniqueString)
-        do { try manager.createParentDirectoryStructure(for: tempArchiveURL) } catch {
-            throw ArchiveError.unwritableArchive }
+        try manager.createParentDirectoryStructure(for: tempArchiveURL)
         guard let tempArchive = Archive(url: tempArchiveURL, accessMode: .create) else {
             throw ArchiveError.unwritableArchive
         }
